split redux actions into separate files
