+++ client/src/components/SignupModal/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { connect } from 'react-redux';
-import { setAuthState } from '../../redux/actions';
+import { setAuthState } from '../../redux/actions/authActions';
 
 import { Button as ReactButton } from 'react-bootstrap';
