+++ client/src/redux/actions/eventDetailActions.js
@@ -1,24 +1,7 @@
-import { SET_AUTH_STATE, DELETE_AUTH_STATE, SET_EVENT } from "../actionTypes";
+import { SET_EVENT } from "../actionTypes";
 
 import API from '../../utils/API';
 
-
-
-/**
- * For logging in or signing up a user in redux
- * @param {} auth 
- */
-export const setAuthState = auth => ({
-  type: SET_AUTH_STATE,
-  payload: auth
-})
-
-/**
- * Remove user details from Redux store
- */
-export const deleteAuthState = () => ({
-  type: DELETE_AUTH_STATE,
-})
 
 
 export const setEvent = (event) => ({
@@ -29,14 +12,13 @@ export const setEvent = (event) => ({
 
 
 export const getEvent = (eventId) => {
-  return function (dispatch) {
+  return async (dispatch) => {
 
-    return API.findEventById(eventId)
-      .then(res => {
-        if (res.data) { // event exists
-          dispatch(setEvent(res.data))
-        }
-      })
+    const res = await API.findEventById(eventId)
+
+    if (res.data) { // event exists
+      return dispatch(setEvent(res.data))
+    }
   }
 }
